+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/spell/Spell.java
@@ -87,6 +87,7 @@ public interface Spell extends SpellTemplate {
     long getProgressLevel();
     boolean cancelOnNoPermission();
     boolean cancelOnNoWand();
+    boolean reactivate();
     @Nonnull
     ConfigurationSection getVariables();
     default void reloadParameters(CastContext context) { };